third_party/blink/web_tests/platform/generic/http/tests/websocket/unmasked-frames-expected.txt

Receive unmasked WebSocket frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
onmessage() was called. (message = "First message")
onmessage() was called. (message = "Fragmented message")
onmessage() was called. (message = "")
onmessage() was called. (message = "END")
PASS actualMessages.length === expectedMessages.length is true
PASS actualMessages[0] is "First message"
PASS actualMessages[1] is "Fragmented message"
PASS actualMessages[2] is ""
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
